requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
py4j==0.10.9.7
PyMySQL==1.1.1
pyspark==3.5.5
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
UNKNOWN @ file:///C:/Users/vscod/OneDrive/Documentos/projects/dba/debito_automatico
urllib3==2.3.0
